added env::ps1 support; move scripts
This work adds env::ps1 support for overriding the hostname and setting custom test in the PS1.  Additionally, this work moves a bunch of config script loading out of the environment system and into profile.d.  Thoe changes will be seen as full removals from the core configuration, but can be re-introduced per-host based on symlinking in the profile.d directory.  This work also adopts the xx-yyyyy pattern for handling order of profile.d script loading.  The env script is set last to ensure all other configuration is loaded before it.  Perhaps 50 would be a better pick in the future though?